LayoutTests/http/tests/websocket/tests/hybi/bad-sub-protocol-non-ascii.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test WebSocket bad sub-protocol names by non-ASCII chars.");

if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

// Fails if protocol contains an character greater than U+007E.
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\u007F")');
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\u0080")');
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\u3042")');
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\uFFFF")');
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\uFEFF")');
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\uFFFE")');
// Surrogate pairs
shouldThrow('new WebSocket("ws://127.0.0.1:8880/simple", "\uD840\uDC0B")');

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
