LayoutTests/http/tests/websocket/tests/hybi/bad-sub-protocol-non-ascii-expected.txt

CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u007F'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u0080'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\u3042'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\uFFFF'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\uFEFF'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\uFFFE'
CONSOLE MESSAGE: line 1: Wrong protocol for WebSocket '\uD840\uDC0B'
Test WebSocket bad sub-protocol names by non-ASCII chars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "あ") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￿") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "﻿") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "￾") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS new WebSocket("ws://127.0.0.1:8880/simple", "𠀋") threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
